+++ src/main/java/org/camunda/community/benchmarks/StartDiExecutor.java
@@ -5,6 +5,7 @@ import io.camunda.zeebe.client.ZeebeClientConfiguration;
 import io.camunda.zeebe.client.api.command.FinalCommandStep;
 import io.camunda.zeebe.client.impl.ZeebeClientImpl;
 import io.camunda.zeebe.spring.client.jobhandling.CommandWrapper;
+import io.camunda.zeebe.spring.client.jobhandling.DefaultCommandExceptionHandlingStrategy;
 import org.camunda.community.benchmarks.config.BenchmarkConfiguration;
 import org.camunda.community.benchmarks.refactoring.RefactoredCommandWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,41 +21,18 @@ import java.util.HashMap;
 import java.util.Map;
 
 @Component
-public class StartDiExecutor {
-
-    public static final String BENCHMARK_START_DATE_MILLIS = "benchmark_di_start_date_millis";
-    private static final Object BENCHMARK_STARTER_ID = "benchmark_di_starter_id";
-
-    @Autowired
-    private BenchmarkConfiguration config;
-
-    @Autowired
-    private ZeebeClient client;
-
-    @Autowired
-    private StatisticsCollector stats;
+public class StartDiExecutor extends StartInstanceExecutor {
 
     @Autowired
     private BenchmarkStartDiExceptionHandlingStrategy exceptionHandlingStrategy;
 
     @Autowired
-    private ZeebeClientConfiguration zeebeClientConfiguration;
+    private StatisticsCollector stats;
 
-    private Map<String, Object> benchmarkPayload;
-
-    @PostConstruct
-    public void init() throws IOException {
-        String variablesJsonString = tryReadVariables(config.getPayloadPath().getInputStream());
-        benchmarkPayload = zeebeClientConfiguration.getJsonMapper().fromJsonAsMap(variablesJsonString);
-    }
-
-    public void startDecisionInstance() {
+    @Override
+    public void startInstance() {
         HashMap<Object, Object> variables = new HashMap<>();
-        variables.putAll(this.benchmarkPayload);
-        variables.put("pincode","515004");
-        variables.put("listItem","c");
-        variables.put(BENCHMARK_START_DATE_MILLIS, Instant.now().toEpochMilli());
-        variables.put(BENCHMARK_STARTER_ID, config.getStarterId());
+        variables.putAll(benchmarkPayload);
 
         FinalCommandStep createCommand = client.newEvaluateDecisionCommand()
                 .decisionId(config.getDmnDecisionId())
@@ -69,16 +47,4 @@ public class StartDiExecutor {
 
     }
 
-    private String tryReadVariables(final InputStream inputStream) throws IOException {
-        final StringBuilder stringBuilder = new StringBuilder();
-        try (final InputStreamReader reader = new InputStreamReader(inputStream)) {
-            try (final BufferedReader br = new BufferedReader(reader)) {
-                String line;
-                while ((line = br.readLine()) != null) {
-                    stringBuilder.append(line).append("\n");
-                }
-            }
-        }
-        return stringBuilder.toString();
-    }
 }
