+++ src/main/resources/bpmn/typical_payload.json
@@ -16,5 +16,7 @@
   "var11": "b18c-9e2145b45074-b2959d57-d091-42d4",
   "var12": 7458,
   "var13": false,
-  "list": ["a", "c"]
+  "var14": "515004",
+  "list": ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j"],
+  "listItem" : "i"
 }
